+++ database/migrations/2017_02_07_162143_add_payment_method_id_to_purchase_order_invoice_table.php
@@ -24,8 +24,11 @@ class AddPaymentMethodIdToPurchaseOrderInvoiceTable extends Migration
      */
     public function down()
     {
-        Schema::table('purchase_order_invoices', function($table){
-            $table->dropColumn('payment_method_id');
-        });
+        if(Schema::hasColumn('purchase_order_invoices', 'payment_method_id')){
+            Schema::table('purchase_order_invoices', function($table){
+                $table->dropColumn('payment_method_id');
+            });    
+        }
+        
     }
 }
